requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.14.0
keras==2.2.5
opencv-python
h5py
imgaug
IPython[all]
pinyin
NLTK
pandas
scikit-learn
djangorestframework


terminaltables
torch==1.8.1
pycocotools
easyocr
mmcv-full==1.3.4 -f https://download.openmmlab.com/mmcv/dist/cpu/torch1.8.1/index.html